ref: move Calendar related components into own dir
